fix-warnings patch
* mpi/mpi.c (mp_to_unsigned_bin, s_mp_mod_2d): Fix signed/unsigned warning.
(mp_toradix): Fix shadowing local variable name warning.

* mpi/mplogic.c (mpl_num_set, mpl_num_clear): Fix signed/unsigned warning.

+++ mpi/mplogic.c
@@ -290,7 +290,7 @@ mp_err mpl_num_set(mp_int *a, int *num)
   for(ix = 0; ix < USED(a); ix++) {
     cur = DIGIT(a, ix);
     
-    for(db = 0; db < sizeof(mp_digit); db++) {
+    for(db = 0; db < (int) sizeof(mp_digit); db++) {
       reg = (cur >> (CHAR_BIT * db)) & UCHAR_MAX;
 
       nset += bitc[reg];
@@ -319,7 +319,7 @@ mp_err mpl_num_clear(mp_int *a, int *num)
   for(ix = 0; ix < USED(a); ix++) {
     cur = DIGIT(a, ix);
     
-    for(db = 0; db < sizeof(mp_digit); db++) {
+    for(db = 0; db < (int) sizeof(mp_digit); db++) {
       reg = (cur >> (CHAR_BIT * db)) & UCHAR_MAX;
 
       nset += bitc[UCHAR_MAX - reg];
